Remove MainDrawingPanel
