--- PixiEditor/Views/UserControls/MainDrawingPanel.xaml
@@ -1,20 +1,0 @@
-﻿<UserControl x:Class="PixiEditor.Views.MainDrawingPanel"
-             xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
-             xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
-             xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006"
-             xmlns:d="http://schemas.microsoft.com/expression/blend/2008"
-             xmlns:local="clr-namespace:PixiEditor.Views"
-             xmlns:userc="clr-namespace:PixiEditor.Views.UserControls"
-             xmlns:i="clr-namespace:System.Windows.Interactivity;assembly=System.Windows.Interactivity"
-             xmlns:helpers="clr-namespace:PixiEditor.Helpers"
-             xmlns:xctk="http://schemas.xceed.com/wpf/xaml/toolkit"
-             mc:Ignorable="d"
-             d:DesignHeight="450" d:DesignWidth="800" x:Name="mainDrawingPanel">
-    <Grid>
-        <userc:Zoombox AdditionalContent="{Binding Item, ElementName=mainDrawingPanel}" x:Name="zoombox" />
-        <xctk:Zoombox Cursor="{Binding Cursor, ElementName=mainDrawingPanel}" KeepContentInBounds="False" Visibility="Hidden"
-                  IsAnimated="False" DragModifiers="Blocked" ZoomModifiers="None">
-
-        </xctk:Zoombox>
-    </Grid>
-</UserControl>
